fix(avo-1724): cleanup flowplayer playlist height + improve styles
